+++ src/main/java/park/bumsiku/service/PublicService.java
@@ -13,7 +13,7 @@ import park.bumsiku.utils.monitoring.LogExecutionTime;
 import park.bumsiku.repository.CategoryRepository;
 import park.bumsiku.repository.CommentRepository;
 import park.bumsiku.repository.PostRepository;
-import park.bumsiku.utils.validation.DiscordWebhookCreator;
+import park.bumsiku.utils.integration.DiscordWebhookCreator;
 
 import java.util.List;
 import java.util.NoSuchElementException;
