+++ src/main/java/park/bumsiku/utils/GlobalExceptionHandler.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
 import park.bumsiku.domain.dto.response.Response;
-import park.bumsiku.utils.validation.DiscordWebhookCreator;
+import park.bumsiku.utils.integration.DiscordWebhookCreator;
 
 import java.util.NoSuchElementException;
