+++ src/main/java/park/bumsiku/utils/integration/DiscordWebhookCreator.java
@@ -1,4 +1,4 @@
-package park.bumsiku.utils.validation;
+package park.bumsiku.utils.integration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
