+++ src/test/java/park/bumsiku/controller/AdminControllerTest.java
@@ -21,8 +21,8 @@ import park.bumsiku.domain.dto.response.CategoryResponse;
 import park.bumsiku.domain.dto.response.PostResponse;
 import park.bumsiku.domain.dto.response.UploadImageResponse;
 import park.bumsiku.service.PrivateService;
-import park.bumsiku.utils.integration.ArgumentValidator;
-import park.bumsiku.utils.validation.DiscordWebhookCreator;
+import park.bumsiku.utils.validation.ArgumentValidator;
+import park.bumsiku.utils.integration.DiscordWebhookCreator;
 
 import java.time.LocalDateTime;
 import java.util.NoSuchElementException;
